fix: resolve 4 clippy errors from airgapped build
- Gate XLMRobertaConfig and serde_json::Value imports under #[cfg(online-model)]
  (unused in airgapped/embedded-model build)
- Fix schema_version parse: u32 -> i64 to match BrainConfig.schema_version field type
- Fix XLMRobertaModel::forward call: pass 4 extra Option args (token_type_ids,
  past_key_values, encoder_hidden_states, encoder_attention_mask)

+++ src/core/db.rs
@@ -284,7 +284,7 @@ pub fn read_brain_config(conn: &Connection) -> Result<Option<BrainConfig>, DbErr
     let schema_version = rows
         .remove("schema_version")
         .unwrap()
-        .parse::<u32>()
+        .parse::<i64>()
         .map_err(|_| DbError::Schema {
             message: "brain_config.schema_version must be an integer".to_owned(),
         })?;
